feat(navbar)!: full migration to WC-backed bs-navbar; delete legacy navbar
BREAKING: the legacy Angular navbar (bs-navbar / bs-navbar-nav / bs-navbar-item
/ bs-navbar-brand / bs-navbar-dropdown / bsNavbarTrigger / bsNavbarContent /
bsExpandButton) is deleted with no back-compat. The WC-backed wrapper is
relocated into the @mintplayer/ng-bootstrap/navbar entry point: bs-navbar /
bs-navbar-brand / bs-navbar-item / bs-navbar-dropdown (classes Bs*Component,
wrapping mp-navbar & friends). ng-package.js shim replaced with ng-package.json
(ng-packagr only discovers .json).

The demo app shell is rewritten to the new API: top-level dropdowns compose
<bs-navbar-dropdown><bs-dropdown-menu><li bsDropdownItem><a routerLink> and
nested submenus place a nested <bs-navbar-dropdown> inside the parent
<bs-dropdown-menu>. The WC owns the toggler/collapse (no bsNavbarToggler /
bs-navbar-nav wrapper) and is position:relative (no fixed-offset padding hack).
navbar-toggler lib kept (used standalone by the toggle-buttons demo). Demo
navbar-wc page folded into the canonical overlays/navbar page. e2e selectors
(bs-navbar host + light-DOM links) still resolve.

nx build mintplayer-ng-bootstrap + nx build ng-bootstrap-demo green; 472 ng
unit tests pass.

Known follow-up: active-route highlighting is not wired per-item on the new
navbar (legacy auto-highlighted the current route).
