+++ apps/ng-bootstrap-demo/src/app/app.component.ts
@@ -1,11 +1,8 @@
 /// <reference types="../types" />
 
 import { Component, inject, ChangeDetectionStrategy } from '@angular/core';
-import { SlideUpDownAnimation } from '@mintplayer/ng-animations';
-import { Color } from '@mintplayer/ng-bootstrap';
-import { BsNavbarComponent, BsNavbarNavComponent, BsNavbarDropdownComponent, BsNavbarItemComponent, BsNavbarContentDirective, BsNavbarBrandComponent, BsExpandButtonDirective, BsNavbarTriggerDirective } from '@mintplayer/ng-bootstrap/navbar';
-import { BsDropdownDividerDirective } from '@mintplayer/ng-bootstrap/dropdown-divider';
-import { BsNavbarTogglerComponent } from '@mintplayer/ng-bootstrap/navbar-toggler';
+import { BsNavbarComponent, BsNavbarDropdownComponent, BsNavbarItemComponent, BsNavbarBrandComponent } from '@mintplayer/ng-bootstrap/navbar';
+import { BsDropdownMenuComponent, BsDropdownItemDirective, BsDropdownDividerDirective } from '@mintplayer/ng-bootstrap/dropdown-menu';
 import { ViewportScroller } from '@angular/common';
 import { RouterLink, RouterOutlet } from '@angular/router';
 import { NavbarRouterLinkActiveDirective } from './directives/navbar-routerlink-active/navbar-router-link-active.directive';
@@ -17,13 +14,11 @@ import { FrameworkLinksComponent } from './components/framework-links/framework-
   selector: 'demo-bootstrap-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.scss'],
-  imports: [RouterOutlet, RouterLink, BsNavbarComponent, BsNavbarNavComponent, BsNavbarDropdownComponent, BsNavbarItemComponent, BsNavbarContentDirective, BsNavbarBrandComponent, BsExpandButtonDirective, BsNavbarTriggerDirective, BsDropdownDividerDirective, BsNavbarTogglerComponent, NavbarRouterLinkActiveDirective, ThemeToggleComponent, FrameworkLinksComponent],
-  animations: [SlideUpDownAnimation],
+  imports: [RouterOutlet, RouterLink, BsNavbarComponent, BsNavbarDropdownComponent, BsNavbarItemComponent, BsNavbarBrandComponent, BsDropdownMenuComponent, BsDropdownItemDirective, BsDropdownDividerDirective, NavbarRouterLinkActiveDirective, ThemeToggleComponent, FrameworkLinksComponent],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class AppComponent {
   versionInfo = inject(BOOTSTRAP_VERSION);
-  colors = Color;
 
   constructor() {
     inject(ViewportScroller).setOffset([0, 56]);
