+++ apps/ng-bootstrap-demo/src/app/pages/overlay/navbar/navbar.component.scss
@@ -1,5 +1,5 @@
 // Frame the live navbar so it reads as a contained example.
-.navbar-wc-demo {
+.navbar-demo {
   background-color: var(--bs-secondary-bg);
   border-radius: 0.5rem;
   padding: 1.5rem;
