+++ apps/ng-bootstrap-demo/src/app/pages/overlay/navbar/navbar.component.ts
@@ -2,32 +2,32 @@ import { ChangeDetectionStrategy, Component, signal } from '@angular/core';
 import { RouterLink } from '@angular/router';
 import { BsCodeSnippetComponent } from '@mintplayer/ng-bootstrap/code-snippet';
 import {
-  BsNavbarWc,
-  BsNavbarItemWc,
-  BsNavbarBrandWc,
-  BsNavbarDropdownWc,
-} from '@mintplayer/ng-bootstrap/navbar-wc';
+  BsNavbarComponent,
+  BsNavbarItemComponent,
+  BsNavbarBrandComponent,
+  BsNavbarDropdownComponent,
+} from '@mintplayer/ng-bootstrap/navbar';
 import { BsDropdownMenuComponent, BsDropdownItemDirective } from '@mintplayer/ng-bootstrap/dropdown-menu';
 import type { NavbarExpandedChangeEventDetail } from '@mintplayer/web-components/navbar';
 import { dedent } from 'ts-dedent';
 
 @Component({
-  selector: 'demo-navbar-wc',
+  selector: 'demo-navbar',
   imports: [
     RouterLink,
     BsCodeSnippetComponent,
-    BsNavbarWc,
-    BsNavbarItemWc,
-    BsNavbarBrandWc,
-    BsNavbarDropdownWc,
+    BsNavbarComponent,
+    BsNavbarItemComponent,
+    BsNavbarBrandComponent,
+    BsNavbarDropdownComponent,
     BsDropdownMenuComponent,
     BsDropdownItemDirective,
   ],
-  templateUrl: './navbar-wc.component.html',
-  styleUrl: './navbar-wc.component.scss',
+  templateUrl: './navbar.component.html',
+  styleUrl: './navbar.component.scss',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class NavbarWcComponent {
+export class NavbarComponent {
   protected readonly expanded = signal(false);
 
   protected onExpandedchange(detail: NavbarExpandedChangeEventDetail) {
@@ -35,41 +35,41 @@ export class NavbarWcComponent {
   }
 
   protected readonly snippetHtml = dedent`
-    <bs-navbar-wc [breakpoint]="'lg'" [color]="'body-tertiary'" [ariaLabel]="'Main navigation'" (expandedchange)="onExpandedchange($event)">
-      <bs-navbar-wc-brand>
+    <bs-navbar [breakpoint]="'lg'" [color]="'body-tertiary'" [ariaLabel]="'Main navigation'" (expandedchange)="onExpandedchange($event)">
+      <bs-navbar-brand>
         <a routerLink="/">MyApp</a>
-      </bs-navbar-wc-brand>
+      </bs-navbar-brand>
 
-      <bs-navbar-wc-item [active]="true">
+      <bs-navbar-item [active]="true">
         <a routerLink="/">Home</a>
-      </bs-navbar-wc-item>
-      <bs-navbar-wc-item>
+      </bs-navbar-item>
+      <bs-navbar-item>
         <a routerLink="/about">About</a>
-      </bs-navbar-wc-item>
+      </bs-navbar-item>
 
-      <bs-navbar-wc-dropdown [label]="'Products'">
+      <bs-navbar-dropdown [label]="'Products'">
         <bs-dropdown-menu>
           <li bsDropdownItem><a routerLink="/products/widgets">Widgets</a></li>
           <li bsDropdownItem><a routerLink="/products/gadgets">Gadgets</a></li>
           <li bsDropdownItem><a routerLink="/products/gizmos">Gizmos</a></li>
         </bs-dropdown-menu>
-      </bs-navbar-wc-dropdown>
+      </bs-navbar-dropdown>
 
-      <bs-navbar-wc-item slot="end">
+      <bs-navbar-item slot="end">
         <a routerLink="/sign-in">Sign in</a>
-      </bs-navbar-wc-item>
-    </bs-navbar-wc>
+      </bs-navbar-item>
+    </bs-navbar>
   `;
 
   protected readonly snippetTs = dedent`
     import { Component, signal } from '@angular/core';
     import { RouterLink } from '@angular/router';
     import {
-      BsNavbarWc,
-      BsNavbarItemWc,
-      BsNavbarBrandWc,
-      BsNavbarDropdownWc,
-    } from '@mintplayer/ng-bootstrap/navbar-wc';
+      BsNavbarComponent,
+      BsNavbarItemComponent,
+      BsNavbarBrandComponent,
+      BsNavbarDropdownComponent,
+    } from '@mintplayer/ng-bootstrap/navbar';
     import { BsDropdownMenuComponent, BsDropdownItemDirective } from '@mintplayer/ng-bootstrap/dropdown-menu';
     import type { NavbarExpandedChangeEventDetail } from '@mintplayer/web-components/navbar';
 
@@ -78,10 +78,10 @@ export class NavbarWcComponent {
       templateUrl: './my-navbar-demo.component.html',
       imports: [
         RouterLink,
-        BsNavbarWc,
-        BsNavbarItemWc,
-        BsNavbarBrandWc,
-        BsNavbarDropdownWc,
+        BsNavbarComponent,
+        BsNavbarItemComponent,
+        BsNavbarBrandComponent,
+        BsNavbarDropdownComponent,
         BsDropdownMenuComponent,
         BsDropdownItemDirective,
       ],
