+++ apps/ng-bootstrap-demo/src/app/pages/overlay/overlay.routes.ts
@@ -12,5 +12,5 @@ export const ROUTES: Routes = [
   { path: 'toast', loadComponent: () => import('./toast/toast.component').then(m => m.ToastComponent) },
   { path: 'shell', loadComponent: () => import('./shell/shell.component').then(m => m.ShellComponent) },
   { path: 'dropdown-wc', loadComponent: () => import('./dropdown-wc/dropdown-wc.component').then(m => m.DropdownWcComponent) },
-  { path: 'navbar-wc', loadComponent: () => import('./navbar-wc/navbar-wc.component').then(m => m.NavbarWcComponent) },
+  { path: 'navbar', loadComponent: () => import('./navbar/navbar.component').then(m => m.NavbarComponent) },
 ];
