--- libs/mintplayer-ng-bootstrap/navbar-wc/index.ts
@@ -1,1 +1,0 @@
-export * from './src';
