--- libs/mintplayer-ng-bootstrap/navbar-wc/ng-package.js
@@ -1,2 +1,0 @@
-// Generated. Shared config + rationale lives in ng-package.secondary.cjs.
-module.exports = require('../ng-package.secondary.cjs').secondaryEntry();
