--- libs/mintplayer-ng-bootstrap/navbar-wc/src/index.ts
@@ -1,4 +1,0 @@
-export * from './navbar/navbar-wc.component';
-export * from './navbar-item/navbar-item-wc.component';
-export * from './navbar-brand/navbar-brand-wc.component';
-export * from './navbar-dropdown/navbar-dropdown-wc.component';
