--- libs/mintplayer-ng-bootstrap/navbar-wc/src/navbar-dropdown/navbar-dropdown-wc.component.html
@@ -1,4 +1,0 @@
-<mp-navbar-dropdown>
-    <span slot="label">{{ label() }}</span>
-    <ng-content></ng-content>
-</mp-navbar-dropdown>
