--- libs/mintplayer-ng-bootstrap/navbar-wc/src/navbar/navbar-wc.component.html
@@ -1,3 +1,0 @@
-<mp-navbar [attr.breakpoint]="breakpoint()" [attr.color]="colorAttr()" [attr.aria-label]="ariaLabelAttr()" [attr.expanded]="expandedAttr()" (expandedchange)="onExpandedchange($event)">
-    <ng-content></ng-content>
-</mp-navbar>
