--- libs/mintplayer-ng-bootstrap/navbar/src/dropdown-toggle/dropdown-toggle.directive.ts
@@ -1,18 +1,0 @@
-import { ContentChildren, Directive, ElementRef, forwardRef, inject, QueryList } from '@angular/core';
-import { BsNavbarItemComponent } from '../navbar-item/navbar-item.component';
-import { BsNavbarDropdownComponent } from '../navbar-dropdown/navbar-dropdown.component';
-
-@Directive({
-  // selector: 'bs-navbar-item > a[routerLink]',
-  selector: 'bs-navbar-item',
-  queries: {
-    childDropdowns: new ContentChildren(forwardRef(() => BsNavbarDropdownComponent))
-  },
-})
-export class DropdownToggleDirective {
-
-  private elementRef = inject<ElementRef<HTMLAnchorElement>>(ElementRef);
-  private bsNavbarItem = inject<BsNavbarItemComponent>(forwardRef(() => BsNavbarItemComponent));
-
-  childDropdowns!: QueryList<BsNavbarDropdownComponent>;
-}
