--- libs/mintplayer-ng-bootstrap/navbar/src/expand-button/expand-button.directive.spec.ts
@@ -1,69 +1,0 @@
-import { Component, signal, ViewChild } from '@angular/core';
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { BsNavbarComponent } from '..';
-import { BsExpandButtonDirective } from './expand-button.directive';
-
-describe('BsExpandButtonDirective', () => {
-  let component: BsExpandButtonTestComponent;
-  let fixture: ComponentFixture<BsExpandButtonTestComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [
-        // Directive to test
-        BsExpandButtonDirective,
-
-        // Mock dependencies
-        BsNavbarMockComponent,
-
-        // Testbench
-        BsExpandButtonTestComponent,
-      ]
-    })
-    .compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(BsExpandButtonTestComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-  
-  it('should put the togglebutton template in the navbar', () => {
-    expect(component.navbar.expandButtonTemplate).toBeDefined();
-  });
-});
-
-@Component({
-  selector: 'bs-navbar',
-  template: `
-    <nav>
-      <div>
-        <ng-content></ng-content>
-      </div>  
-    </nav>`,
-  providers: [
-    { provide: BsNavbarComponent, useExisting: BsNavbarMockComponent }
-  ]
-})
-class BsNavbarMockComponent {
-  expandButtonTemplate = signal<any>(null);
-}
-
-@Component({
-  selector: 'bs-nav-link-test',
-  imports: [BsNavbarMockComponent, BsExpandButtonDirective],
-  template: `
-    <bs-navbar #navbar>
-      <ng-template bsExpandButton let-state>
-        <input type="checkbox" [checked]="state === 'open'">
-      </ng-template>
-    </bs-navbar>`
-})
-class BsExpandButtonTestComponent {
-  @ViewChild('navbar') navbar!: BsNavbarComponent;
-}
