+++ libs/mintplayer-ng-bootstrap/navbar/src/index.ts
@@ -1,12 +1,4 @@
 export * from './navbar/navbar.component';
-export * from './navbar-brand/navbar-brand.component';
-export * from './navbar-nav/navbar-nav.component';
-export * from './navbar-dropdown/navbar-dropdown.component';
 export * from './navbar-item/navbar-item.component';
-export * from './navbar-content/navbar-content.directive';
-export * from './navbar-trigger/navbar-trigger.directive';
-export * from './expand-button/expand-button.directive';
-
-// Not for end-use
-export * from './dropdown-toggle/dropdown-toggle.directive';
-export * from './nav-link/nav-link.directive';
+export * from './navbar-brand/navbar-brand.component';
+export * from './navbar-dropdown/navbar-dropdown.component';
