--- libs/mintplayer-ng-bootstrap/navbar/src/nav-link/nav-link.directive.spec.ts
@@ -1,144 +1,0 @@
-import { Component } from '@angular/core';
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { RouterLink } from '@angular/router';
-import { RouterTestingModule } from '@angular/router/testing';
-import { MockComponent } from 'ng-mocks';
-import { BsNavbarComponent } from '../navbar/navbar.component';
-import { BsNavbarNavComponent } from '../navbar-nav/navbar-nav.component';
-import { BsNavbarItemComponent } from '../navbar-item/navbar-item.component';
-import { BsNavbarDropdownComponent } from '../navbar-dropdown/navbar-dropdown.component';
-import { NavLinkDirective } from './nav-link.directive';
-
-describe('NavLinkDirective', () => {
-  let component: BsNavLinkTestComponent;
-  let fixture: ComponentFixture<BsNavLinkTestComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [
-        RouterTestingModule.withRoutes([
-          { path: 'a', component: PageAComponent },
-          { path: 'b', children: [
-            { path: 'c', component: PageBCComponent }
-          ]}
-        ]),
-        // Directive to test
-        NavLinkDirective,
-
-        // Mock components
-        BsNavbarMockComponent,
-        BsNavbarNavMockComponent,
-        BsNavbarDropdownMockComponent,
-        BsNavbarItemMockComponent,
-
-        // Pages
-        PageAComponent,
-        PageBCComponent,
-
-        // Testbench
-        BsNavLinkTestComponent,
-      ],
-      providers: [
-        { provide: BsNavbarDropdownComponent, useClass: BsNavbarDropdownMockComponent }
-      ]
-    })
-    .compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(BsNavLinkTestComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
-
-@Component({
-  selector: 'bs-nav-link-test',
-  template: `
-    <bs-navbar>
-      <bs-navbar-nav>
-        <bs-navbar-item>
-          <a [routerLink]='["/a"]'>a</a>
-        </bs-navbar-item>
-        <bs-navbar-item>
-          <a [routerLink]='[]'>dropdown</a>
-          <bs-navbar-dropdown>
-            <bs-navbar-item>
-              <a [routerLink]='["/b", "c"]'>bc</a>
-            </bs-navbar-item>
-          </bs-navbar-dropdown>
-        </bs-navbar-item>
-      </bs-navbar-nav>
-    </bs-navbar>`,
-  imports: [
-    RouterLink,
-    MockComponent(BsNavbarComponent),
-    MockComponent(BsNavbarNavComponent),
-    MockComponent(BsNavbarItemComponent),
-    MockComponent(BsNavbarDropdownComponent),
-  ],
-})
-class BsNavLinkTestComponent {
-}
-
-@Component({
-  selector: 'bs-navbar',
-  template: `
-    <nav>
-      <div>
-        <ng-content></ng-content>
-      </div>  
-    </nav>`
-})
-class BsNavbarMockComponent {
-}
-
-@Component({
-  selector: 'bs-navbar-nav',
-  template: `
-    <div>
-      <ul>
-        <ng-content></ng-content>
-      </ul>  
-    </div>`
-})
-class BsNavbarNavMockComponent {
-}
-
-@Component({
-  selector: 'bs-navbar-dropdown',
-  template: `
-    <ul>
-      <ng-content></ng-content>
-    </ul>`
-})
-class BsNavbarDropdownMockComponent {
-}
-
-@Component({
-  selector: 'bs-navbar-item',
-  template: `
-    <li>
-      <ng-content></ng-content>
-    </li>`
-})
-class BsNavbarItemMockComponent {
-}
-
-@Component({
-  selector: 'page-a',
-  template: `<div>Page A</div>`
-})
-class PageAComponent {
-}
-
-@Component({
-  selector: 'page-bc',
-  template: `<div>Page B - C</div>`
-})
-class PageBCComponent {
-}
