--- libs/mintplayer-ng-bootstrap/navbar/src/navbar-brand/navbar-brand.component.html
@@ -1,3 +1,0 @@
-<span class="navbar-brand mx-auto mx-md-unset">
-    <ng-content></ng-content>
-</span>
