--- libs/mintplayer-ng-bootstrap/navbar/src/navbar-brand/navbar-brand.component.spec.ts
@@ -1,29 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { RouterTestingModule } from '@angular/router/testing';
-
-import { BsNavbarBrandComponent } from './navbar-brand.component';
-
-describe('BsNavbarBrandComponent', () => {
-  let component: BsNavbarBrandComponent;
-  let fixture: ComponentFixture<BsNavbarBrandComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [
-        RouterTestingModule.withRoutes([]),
-        BsNavbarBrandComponent
-      ],
-    })
-    .compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(BsNavbarBrandComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
