+++ libs/mintplayer-ng-bootstrap/navbar/src/navbar-dropdown/navbar-dropdown.component.html
@@ -1,14 +1,4 @@
-<bs-has-overlay></bs-has-overlay>
-<ul #dd class="dropdown-menu me-3" [class.show]="showClass()" [class.submenu]="!!parentDropdown" [class.overflow-y-auto]="isBrowser" [style.max-height]="maxHeightOrNull()" bsNoNoscript (clickOutside)="isVisible.set(false)" [exclude]="elementsToExclude" aria-labelledby="navbarDropdown">
+<mp-navbar-dropdown>
+    <span slot="label">{{ label() }}</span>
     <ng-content></ng-content>
-</ul>
-
-<!--
-    - Working demo: https://codepen.io/agop/pen/VwwMGR
-    - Location must be set through javascript
-    - bsNoScript => If noscript, 2nd-level submenus must be shown inside the 1st-level submenu
- -->
- <!--
-    - Another option: https://stackblitz.com/edit/web-platform-vmsr4l?file=index.html
-    - But this doesn't seem to work here
-  -->
+</mp-navbar-dropdown>
