--- libs/mintplayer-ng-bootstrap/navbar/src/navbar-dropdown/navbar-dropdown.component.spec.ts
@@ -1,161 +1,0 @@
-import { Component, computed, Directive, ElementRef, Input, signal } from '@angular/core';
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { RouterLink } from '@angular/router';
-import { RouterTestingModule } from '@angular/router/testing';
-import { BsHasOverlayComponent } from '@mintplayer/ng-bootstrap/has-overlay';
-import { MockComponent } from 'ng-mocks';
-import { BsNavbarItemComponent } from '../navbar-item/navbar-item.component';
-import { BsNavbarComponent } from '../navbar/navbar.component';
-
-import { BsNavbarDropdownComponent } from './navbar-dropdown.component';
-
-describe('BsNavbarDropdownComponent', () => {
-  let component: BsNavbarDropdownTestComponent;
-  let fixture: ComponentFixture<BsNavbarDropdownTestComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [
-        RouterTestingModule.withRoutes([
-          { path: 'a', component: PageAComponent },
-          { path: 'b', children: [
-            { path: 'c', component: PageBCComponent }
-          ]}
-        ]),
-        MockComponent(BsHasOverlayComponent),
-        // Component to test
-        BsNavbarDropdownComponent,
-
-        // Mock components
-        BsNavbarMockComponent,
-        BsNavbarNavMockComponent,
-        BsNavbarItemMockComponent,
-        ClickOutsideDirective,
-
-        // Pages
-        PageAComponent,
-        PageBCComponent,
-
-        // Testbench
-        BsNavbarDropdownTestComponent
-      ],
-    })
-    .compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(BsNavbarDropdownTestComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
-
-@Component({
-  selector: 'bs-navbar',
-  template: `
-    <nav>
-      <div>
-        <ng-content></ng-content>
-      </div>
-    </nav>`,
-  providers: [
-    { provide: BsNavbarComponent, useExisting: BsNavbarMockComponent }
-  ]
-})
-class BsNavbarMockComponent {
-  isExpanded = signal<boolean>(false);
-  windowWidth = signal<number | null>(1024);
-  breakpoint = signal<string | null>('md');
-  expandAt = computed(() => {
-    switch (this.breakpoint()) {
-      case 'xxl': return 1400;
-      case 'xl': return 1200;
-      case 'lg': return 992;
-      case 'md': return 768;
-      case 'sm': return 576;
-      case 'xs': return 0;
-      default: return null;
-    }
-  });
-  isSmallMode = signal<boolean>(false);
-}
-
-@Component({
-  selector: 'bs-navbar-nav',
-  template: `
-  <div>
-    <ul>
-      <ng-content></ng-content>
-    </ul>
-  </div>`
-})
-class BsNavbarNavMockComponent {
-}
-
-@Component({
-  selector: 'bs-navbar-item',
-  template: `
-  <li>
-    <ng-content></ng-content>
-  </li>`,
-  providers: [
-    { provide: BsNavbarItemComponent, useExisting: BsNavbarItemMockComponent }
-  ]
-})
-class BsNavbarItemMockComponent {
-  element = { nativeElement: document.createElement('div') };
-}
-
-@Component({
-  selector: 'bs-navbar-test',
-  template: `
-  <bs-navbar>
-    <bs-navbar-nav>
-      <bs-navbar-item>
-        <a [routerLink]='["/a"]'>a</a>
-      </bs-navbar-item>
-      <bs-navbar-item>
-        <a [routerLink]='[]'>dropdown</a>
-        <bs-navbar-dropdown>
-          <bs-navbar-item>
-            <a [routerLink]='["/b", "c"]'>bc</a>
-          </bs-navbar-item>
-        </bs-navbar-dropdown>
-      </bs-navbar-item>
-    </bs-navbar-nav>
-  </bs-navbar>`,
-  imports: [
-    RouterLink,
-    BsNavbarMockComponent,
-    BsNavbarNavMockComponent,
-    BsNavbarItemMockComponent,
-    BsNavbarDropdownComponent,
-  ],
-})
-class BsNavbarDropdownTestComponent {
-}
-
-@Component({
-  selector: 'page-a',
-  template: `<div>Page A</div>`
-})
-class PageAComponent {
-}
-
-@Component({
-  selector: 'page-bc',
-  template: `<div>Page B - C</div>`
-})
-class PageBCComponent {
-}
-
-@Directive({
-  selector: '[clickOutside]',
-})
-class ClickOutsideDirective {
-  @Input() exclude: HTMLElement[] = [];
-}
