--- libs/mintplayer-ng-bootstrap/navbar/src/navbar-item/navbar-item.component.html
@@ -1,3 +1,0 @@
-<li [class.nav-item]="parentDropdown === null" [class.dropdown]="(dropdowns().length > 0) && (parentDropdown === null)" [class.dropend]="(dropdowns().length > 0) && (parentDropdown !== null)">
-    <ng-content></ng-content>
-</li>
