--- libs/mintplayer-ng-bootstrap/navbar/src/navbar-item/navbar-item.component.scss
@@ -1,9 +1,0 @@
-li.dropend ::ng-deep a.dropdown-toggle {
-    padding-right: 1.5rem;
-    
-    &::after {
-        position: absolute;
-        right: .5rem;
-        top: .75rem;
-    }
-}
