--- libs/mintplayer-ng-bootstrap/navbar/src/navbar-nav/navbar-nav.component.html
@@ -1,5 +1,0 @@
-<div class="navbar-collapse w-100 show" [class.collapse]="collapse()" [@slideUpDown]="showNavs()" [@.disabled]="isResizing()">
-    <ul class="navbar-nav mr-auto">
-        <ng-content></ng-content>
-    </ul>
-</div>
