--- libs/mintplayer-ng-bootstrap/navbar/src/navbar-nav/navbar-nav.component.scss
@@ -1,3 +1,0 @@
-:host {
-    margin-right: auto;
-}
