--- libs/mintplayer-ng-bootstrap/navbar/src/navbar-nav/navbar-nav.component.spec.ts
@@ -1,140 +1,0 @@
-import { Component, computed, signal } from '@angular/core';
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { RouterLink } from '@angular/router';
-import { RouterTestingModule } from '@angular/router/testing';
-import { NoopAnimationsModule } from '@angular/platform-browser/animations';
-import { BsNavbarComponent } from '../navbar/navbar.component';
-
-import { BsNavbarNavComponent } from './navbar-nav.component';
-
-describe('BsNavbarNavComponent', () => {
-  let component: BsNavbarNavTestComponent;
-  let fixture: ComponentFixture<BsNavbarNavTestComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [
-        NoopAnimationsModule,
-        RouterTestingModule.withRoutes([
-          { path: 'a', component: PageAComponent },
-          { path: 'b', children: [
-            { path: 'c', component: PageBCComponent }
-          ]}
-        ]),
-
-        // Pages
-        PageAComponent,
-        PageBCComponent,
-        // Component to test
-        BsNavbarNavComponent,
-
-        // Mock components
-        BsNavbarMockComponent,
-        BsNavbarDropdownMockComponent,
-        BsNavbarItemMockComponent,
-
-        // Testbench
-        BsNavbarNavTestComponent
-      ],
-      providers: [
-        { provide: BsNavbarComponent, useClass: BsNavbarMockComponent }
-      ]
-    })
-    .compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(BsNavbarNavTestComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
-
-@Component({
-  selector: 'bs-navbar',
-  template: `
-    <nav>
-      <div>
-        <ng-content></ng-content>
-      </div>
-    </nav>`
-})
-class BsNavbarMockComponent {
-  isExpanded = signal<boolean>(false);
-  breakpoint = signal<string | null>('md');
-  expandAt = computed(() => {
-    switch (this.breakpoint()) {
-      case 'xxl': return 1400;
-      case 'xl': return 1200;
-      case 'lg': return 992;
-      case 'md': return 768;
-      case 'sm': return 576;
-      case 'xs': return 0;
-      default: return null;
-    }
-  });
-}
-
-@Component({
-  selector: 'bs-navbar-dropdown',
-  template: `
-    <ul>
-      <ng-content></ng-content>
-    </ul>`
-})
-class BsNavbarDropdownMockComponent {
-}
-
-@Component({
-  selector: 'bs-navbar-item',
-  template: `
-    <li>
-      <ng-content></ng-content>
-    </li>`
-})
-class BsNavbarItemMockComponent {}
-
-@Component({
-  selector: 'bs-navbar-test',
-  template: `
-  <bs-navbar>
-    <bs-navbar-nav>
-      <bs-navbar-item>
-        <a [routerLink]='["/a"]'>a</a>
-      </bs-navbar-item>
-      <bs-navbar-item>
-        <a [routerLink]='[]'>dropdown</a>
-        <bs-navbar-dropdown>
-          <bs-navbar-item>
-            <a [routerLink]='["/b", "c"]'>bc</a>
-          </bs-navbar-item>
-        </bs-navbar-dropdown>
-      </bs-navbar-item>
-    </bs-navbar-nav>
-  </bs-navbar>`,
-  imports: [
-    RouterLink,
-    BsNavbarMockComponent,
-    BsNavbarNavComponent,
-    BsNavbarItemMockComponent,
-    BsNavbarDropdownMockComponent,
-  ],
-})
-class BsNavbarNavTestComponent {
-}
-
-@Component({
-  selector: 'page-a',
-  template: `<div>Page A</div>`
-})
-class PageAComponent {}
-
-@Component({
-  selector: 'page-bc',
-  template: `<div>Page B - C</div>`
-})
-class PageBCComponent {}
