+++ libs/mintplayer-ng-bootstrap/navbar/src/navbar/navbar.component.html
@@ -1,66 +1,3 @@
-<bs-container>
-    <nav #nav class="navbar position-fixed" [class]="navClassList()" [attr.data-bs-theme]="dataBsTheme()" bsUserAgent bsNoNoscript [attr.aria-label]="ariaLabel()">
-        <div class="container-fluid nav-wrapper">
-            <div class="d-flex w-100" [class]="wAutoClass()">
-                <ng-content select="bs-navbar-brand"></ng-content>
-                <div class="flex-end d-flex tw" [class]="dNoneClass()">
-                    @if (isServerSide) {
-                        <!--
-                            SSR-only toggler: hidden checkbox + visible label.
-                            Works as a CSS-only toggle in noscript because:
-                            - The label[for=…] toggles the checkbox on click,
-                              regardless of browser focus-on-click behavior
-                              (Safari + macOS Chrome don't focus buttons on
-                              click, so a plain :focus-within rule fails).
-                            - The input is focusable (Tab + Space toggles it),
-                              so keyboard users get the same behavior.
-                            - The checkbox's :checked state drives a CSS rule
-                              on .navbar.noscript that overrides the SSR-baked
-                              height:0 the @slideUpDown void state emits.
-                            In the browser, isServerSide is false even during
-                            hydration so the @else branch's plain <button>
-                            renders and the JS click handler does the work.
-                        -->
-                        <input type="checkbox"
-                            [id]="togglerCheckboxId"
-                            class="navbar-toggler-checkbox visually-hidden"
-                            [checked]="isExpanded()"
-                            aria-label="Toggle navigation"
-                            [attr.aria-expanded]="isExpanded()"
-                            [attr.aria-controls]="collapseId()" />
-                        <label
-                            [attr.for]="togglerCheckboxId"
-                            class="align-self-center cursor-pointer mb-0"
-                            [class.navbar-toggler]="!expandButtonTemplate()"
-                            [class.bg-transparent]="expandButtonTemplate()"
-                            [class.shadow-none]="expandButtonTemplate()"
-                            [class.border-0]="expandButtonTemplate()"
-                            aria-hidden="true">
-                            <ng-container *ngTemplateOutlet="expandButtonTemplate() ?? defaultExpandButton; context: { $implicit: isExpanded() }"></ng-container>
-                        </label>
-                    } @else {
-                        <button (click)="toggleExpanded()"
-                            class="align-self-center"
-                            [class.navbar-toggler]="!expandButtonTemplate()"
-                            [class.bg-transparent]="expandButtonTemplate()"
-                            [class.shadow-none]="expandButtonTemplate()"
-                            [class.border-0]="expandButtonTemplate()"
-                            type="button"
-                            aria-label="Toggle navigation"
-                            [attr.aria-expanded]="isExpanded()"
-                            [attr.aria-controls]="collapseId()">
-                            <ng-container *ngTemplateOutlet="expandButtonTemplate() ?? defaultExpandButton; context: { $implicit: isExpanded() }"></ng-container>
-                        </button>
-                    }
-                </div>
-            </div>
-            <div [id]="collapseId()" style="display: contents;">
-                <ng-content></ng-content>
-            </div>
-        </div>
-    </nav>
-</bs-container>
-
-<ng-template #defaultExpandButton>
-    <span class="navbar-toggler-icon" aria-hidden="true"></span>
-</ng-template>
+<mp-navbar [attr.breakpoint]="breakpoint()" [attr.color]="colorAttr()" [attr.aria-label]="ariaLabelAttr()" [attr.expanded]="expandedAttr()" (expandedchange)="onExpandedchange($event)">
+    <ng-content></ng-content>
+</mp-navbar>
