--- libs/mintplayer-ng-bootstrap/navbar/src/navbar/navbar.component.scss
@@ -1,132 +1,0 @@
-// Configuration
-@import "node_modules/bootstrap/scss/functions";
-@import "node_modules/bootstrap/scss/variables";
-@import "node_modules/bootstrap/scss/variables-dark";
-@import "node_modules/bootstrap/scss/mixins";
-
-/* :host */ ::ng-deep {
-    // Layout & components
-    // @import "node_modules/bootstrap/scss/containers";
-    @import "node_modules/bootstrap/scss/dropdown";
-    @import "node_modules/bootstrap/scss/nav";
-    @import "node_modules/bootstrap/scss/navbar";
-}
-
-::ng-deep {
-    .navbar {
-        z-index: 20 !important;
-
-        &.noscript {
-            // See docs/prd/navbar-noscript.md for the full contract.
-            //
-            // SSR bakes `style="height: 0"` into .navbar-collapse via
-            // `[@slideUpDown]="false"` (windowWidth is null on the server, so
-            // showNavs() resolves to false). In noscript that inline style
-            // sticks and hides the menu until something overrides it. Three
-            // override paths, each !important to beat the inline style:
-            //
-            //  - Wide mode (viewport ≥ navbar-expand-{bp}): always reveal.
-            //    Bootstrap's own media query gives us `display: flex` on the
-            //    collapse but leaves height to its JS plugin — that's what
-            //    the @each loop below fills in.
-            //  - :has(.navbar-toggler-checkbox:checked) — mouse/touch user
-            //    clicked the hamburger label. We use a label-for-checkbox
-            //    rather than :focus-within on a <button> because Safari and
-            //    macOS Chrome don't focus buttons on mouse click.
-            //  - :focus-within — keyboard user Tabbed into the hidden
-            //    checkbox (or a menu item, once revealed).
-            bs-navbar-nav .navbar-collapse {
-                overflow: hidden;
-            }
-
-            &:has(.navbar-toggler-checkbox:checked) bs-navbar-nav .navbar-collapse,
-            &:focus-within bs-navbar-nav .navbar-collapse {
-                height: auto !important;
-                overflow: visible;
-            }
-
-            @each $breakpoint, $min in $grid-breakpoints {
-                @if $min and $min > 0 {
-                    @include media-breakpoint-up($breakpoint, $grid-breakpoints) {
-                        &.navbar-expand-#{$breakpoint} bs-navbar-nav .navbar-collapse {
-                            height: auto !important;
-                            overflow: visible;
-                        }
-                    }
-                }
-            }
-
-            li.nav-item:focus-within > bs-navbar-dropdown > ul {
-                display: block;
-
-                li:has(a.dropdown-item):focus-within > bs-navbar-dropdown > ul {
-                    display: block;
-                }
-            }
-        }
-    }
-}
-
-nav {
-    left: 0;
-    top: 0;
-    right: 0;
-    z-index: 10;
-    // overflow-y: auto !important;
-}
-
-@media (min-width: 768px) {
-    .mx-md-unset {
-        margin-left: 0 !important;    
-        margin-right: 0 !important;    
-    }
-
-    .w-md-auto {
-        width: auto !important;
-    }
-}
-
-.tw {
-    margin: -.5rem 0;
-}
-
-::ng-deep bs-navbar-nav {
-    display: contents;
-}
-
-// Prevent navbar from going outside its bounds in small mode
-@each $breakpoint, $container-max-width in $container-max-widths {
-    @include media-breakpoint-down($breakpoint, $grid-breakpoints) {
-        nav.navbar-expand-#{$breakpoint} {
-            ::ng-deep div.navbar-collapse {
-                overflow-y: hidden;
-            }
-
-            // Prevent dropdown-menu from exceeding navbar width in small mode
-            ::ng-deep .dropdown-menu {
-                min-width: 0;
-                max-width: 100%;
-            }
-
-            &.position-fixed {
-                max-height: 100%;
-                overflow-y: scroll;
-            }
-
-            &.os-windows {
-                margin-right: -17px;
-            }
-        }
-    }
-}
-
-// Properly align the navbar-navs
-.nav-wrapper ::ng-deep bs-navbar-nav > .navbar-collapse {
-    justify-content: center;
-}
-.nav-wrapper ::ng-deep bs-navbar-nav:first-of-type > .navbar-collapse {
-    justify-content: flex-start;
-}
-.nav-wrapper ::ng-deep bs-navbar-nav:last-of-type > .navbar-collapse {
-    justify-content: flex-end;
-}
